requirements (2).txt
streamlit
plotly
numpy
google-generativeai